refactor: added app to cmd package, made the necessary changes inside the source code for that
