+++ pkg/rpaas/client/go.mod
@@ -2,8 +2,4 @@ module github.com/tsuru/rpaas-operator/pkg/rpaas/client
 
 go 1.13
 
-require (
-	github.com/mattn/go-runewidth v0.0.4 // indirect
-	github.com/olekukonko/tablewriter v0.0.1
-	github.com/stretchr/testify v1.4.0
-)
+require github.com/stretchr/testify v1.4.0
